backend/requirements.txt
# Flask and extensions
Flask==2.2.3
Flask-Cors==3.0.10
Flask-SocketIO==5.3.3

# Database
psycopg2-binary==2.9.5
SQLAlchemy==2.0.5

# Environment variables
python-dotenv==1.0.0

# Data processing
pandas==1.5.3
numpy==1.24.2

# Biometric device integration
git+https://github.com/fananimi/pyzk.git#egg=pyzk

# Utilities
requests==2.28.2
python-dateutil==2.8.2
tabulate==0.9.0
colorama==0.4.6

# Testing
pytest==7.3.1